@@ -1,23 +1,21 @@
 package gamble.card;
 
-import gamble.card.Card;
-import gamble.card.CardEventListener;
-import gamble.card.CardInputter;
-import gamble.card.CardService;
 import gamble.player.Player;
-import org.junit.Before;
-import org.junit.Test;
-import org.junit.runner.RunWith;
+import org.junit.jupiter.api.BeforeEach;
+import org.junit.jupiter.api.Test;
+import org.junit.jupiter.api.extension.ExtendWith;
 import org.mockito.InjectMocks;
 import org.mockito.Mock;
 import org.mockito.junit.MockitoJUnitRunner;
+import org.mockito.junit.jupiter.MockitoExtension;
 
 import java.util.ArrayList;
 
-import static org.junit.Assert.assertEquals;
+import static org.hamcrest.MatcherAssert.assertThat;
+import static org.hamcrest.Matchers.equalTo;
 import static org.mockito.Mockito.when;
 
-@RunWith(MockitoJUnitRunner.class)
+@ExtendWith(MockitoExtension.class)
 public class CardServiceTest {
 
   @Mock
@@ -30,48 +28,48 @@ public class CardServiceTest {
   CardService cs;
 
   // Test data
-  private Player p = new Player();
+  private final Player p = new Player();
 
-  @Before
-  public void setup() {
+  @BeforeEach
+  void setup() {
 
     // Set up test data
     Card card1 = new Card(1, 3, 3);
     Card card2 = new Card(2, 10, 3);
 
-    p.cards = new ArrayList<>();
-
-    p.cards.add(card1);
-    p.cards.add(card2);
+    var cards = new ArrayList<Card>();
+    cards.add(card1);
+    cards.add(card2);
+    p.setCards(cards);
   }
 
   @Test
-  public void rechargeCard1() {
+  void rechargeCard1() {
 
     // Given
-    p.cards.get(0).uses--;
-    when(inMock.selectCard(p.cards)).thenReturn(p.cards.get(0));
+    p.getCards().get(0).uses--;
+    when(inMock.selectCard(p.getCards())).thenReturn(p.getCards().get(0));
 
     // When
     cs.rechargeCard(p);
 
     // Then
-    assertEquals(p.cards.get(0).maxUses, p.cards.get(0).uses);
+    assertThat(p.getCards().get(0).maxUses, equalTo(p.getCards().get(0).uses));
   }
 
   @Test
-  public void rechargeCard2Only() {
+  void rechargeCard2Only() {
 
     // Given
-    p.cards.get(0).uses--;
-    p.cards.get(1).uses--;
-    when(inMock.selectCard(p.cards)).thenReturn(p.cards.get(1));
+    p.getCards().get(0).uses--;
+    p.getCards().get(1).uses--;
+    when(inMock.selectCard(p.getCards())).thenReturn(p.getCards().get(1));
 
     // When
     cs.rechargeCard(p);
 
     // Then
-    assertEquals(p.cards.get(1).maxUses, p.cards.get(1).uses);
-    assertEquals(2, p.cards.get(0).uses);
+    assertThat(p.getCards().get(1).maxUses, equalTo(p.getCards().get(1).uses));
+    assertThat(2, equalTo(p.getCards().get(0).uses));
   }
 }
