@@ -3,12 +3,17 @@ package gamble;
 import java.util.Random;
 
 public class Util {
+  private static final Random r = new Random();
+
+  private Util() {
+    throw new IllegalStateException("Not intended to be instantiated");
+  }
+
   public static void pause(long waitTime) {
     try {
       Thread.sleep(waitTime);
     } catch (InterruptedException e) {
-      // TODO Auto-generated catch block
-      e.printStackTrace();
+      throw new RuntimeException(e);
     }
   }
 
@@ -35,10 +40,9 @@ public class Util {
    * Produces a random value between min and max
    * @param min Minimum value
    * @param max Maximum value
-   * @return
+   * @return Random value
    */
   public static int randomBetween(int min, int max) {
-    Random r = new Random();
     int random = r.nextInt(max - min) + 1;
     return min + random;
   }
