@@ -1,12 +1,11 @@
 package gamble.fight;
 
-public class FightResult {
-  public boolean exactHit = false;
-  public boolean won = false;
+import lombok.Getter;
+import lombok.RequiredArgsConstructor;
 
-  public FightResult(boolean exactHit, boolean won) {
-    super();
-    this.exactHit = exactHit;
-    this.won = won;
-  }
+@RequiredArgsConstructor
+@Getter
+public class FightResult {
+  private final boolean exactHit;
+  private final boolean won;
 }
