@@ -1,6 +1,11 @@
 package gamble.fight;
 
+import lombok.Getter;
+import lombok.Setter;
+
+@Getter
+@Setter
 public class Fighter {
-  public int damageTaken;
-  public int maxHp;
+  private int damageTaken;
+  private int maxHp;
 }
