@@ -5,29 +5,24 @@ import gamble.Util;
 import gamble.card.CardConsoleOutputter;
 import gamble.card.CardService;
 import gamble.player.Player;
+import lombok.RequiredArgsConstructor;
 
+@RequiredArgsConstructor
 public class MatchConsoleOutputter implements MatchEventListener {
   private final CardConsoleOutputter co;
-
-  public MatchConsoleOutputter(CardService cs, CardConsoleOutputter co) {
-    super();
-    this.co = co;
-  }
 
   /* (non-Javadoc)
    * @see gamble.service.output.MatchOutputter#gameOver()
    */
   @Override
   public void matchLost() {
-    String str = String.format("Uh oh, you're out of cards :(");
-    Util.print(str);
+    Util.print("Uh oh, you're out of cards :(");
   }
 
   @Override
   public void exactHit(Player p) {
-    String str = "You got a precise hit! Choose a card to recharge: ";
-    Util.print(str);
-    co.showPlayerCards(p.cards);
+    Util.print("You got a precise hit! Choose a card to recharge: ");
+    co.showPlayerCards(p.getCards());
   }
 
   @Override
