@@ -6,6 +6,7 @@ import gamble.fight.FightService;
 import gamble.fight.Fighter;
 import gamble.fight.FightResult;
 import gamble.player.Player;
+import lombok.RequiredArgsConstructor;
 
 import java.util.ArrayList;
 import java.util.List;
@@ -13,39 +14,40 @@ import java.util.List;
 /**
  * Plays a match against the CPU. A match has many rounds, one for each CPU card.
  */
+@RequiredArgsConstructor
 public class MatchService {
-  private final List<MatchEventListener> matchEventListeners = new ArrayList<MatchEventListener>();
+  private final List<MatchEventListener> eventListeners = new ArrayList<>();
   private final FightService fightService;
   private final CardService cardService;
 
-  public MatchService(FightService fightService, CardService cardService) {
-    this.fightService = fightService;
-    this.cardService = cardService;
-  }
-
-  public MatchResult play(Player p, int cpuCards) {
+  /**
+   * Play a match against a clan of fighters
+   * @param p Player
+   * @param clanFighterCount Number of fighters in the clan
+   * @return Result of the match
+   */
+  public MatchResult play(Player p, int clanFighterCount) {
     int roundNum = 1;
 
-    for (; cpuCards > 0; cpuCards--) {
-
+    for (; clanFighterCount > 0; clanFighterCount--) {
       int finalRoundNum = roundNum;
-      int finalCpuCards = cpuCards;
-      matchEventListeners.forEach(mel -> mel.roundStarted(finalRoundNum, finalCpuCards));
+      int finalCpuCards = clanFighterCount;
+      eventListeners.forEach(e -> e.roundStarted(finalRoundNum, finalCpuCards));
       Fighter fighter = fightService.createFighter(
         Config.MIN_CPU_CARD_VALUE,
         Config.MAX_CPU_CARD_VALUE);
 
       FightResult rr = fightService.fight(fighter, p);
 
-      if (!rr.won) {
-        matchEventListeners.forEach(mel -> mel.matchLost());
+      if (!rr.isWon()) {
+        eventListeners.forEach(MatchEventListener::matchLost);
         return new MatchResult(false);
       } else {
         roundNum++;
       }
 
-      if (rr.exactHit) {
-        matchEventListeners.forEach(mel -> mel.exactHit(p));
+      if (rr.isExactHit()) {
+        eventListeners.forEach(mel -> mel.exactHit(p));
         cardService.rechargeCard(p);
       }
     }
@@ -53,6 +55,6 @@ public class MatchService {
   }
 
   public void addMatchEventListener(MatchEventListener matchEventListener) {
-    matchEventListeners.add(matchEventListener);
+    eventListeners.add(matchEventListener);
   }
 }
